Move region functions from region endpoint to bcgw-layer-service.ts
